added all buttons to interact with minus iframe (some missing functionality to be added)
