+++ client/components/App/queries.js
@@ -1,0 +1,11 @@
+import { gql } from '@apollo/client';
+
+export const ME_QUERY = gql`
+    query {
+        me {
+            id
+            username
+            roles
+        }
+    }
+`;
