+++ client/components/TestQueue/queries.js
@@ -1,15 +1,5 @@
 import { gql } from '@apollo/client';
 
-export const ME_QUERY = gql`
-    query {
-        me {
-            id
-            username
-            roles
-        }
-    }
-`;
-
 export const TEST_QUEUE_PAGE_QUERY = gql`
     query {
         users {
