+++ server/resolvers/TestPlanRun/index.js
@@ -1,7 +1,9 @@
+const isComplete = require('./isCompleteResolver');
 const testResultCount = require('./testResultCountResolver');
 const testResults = require('./testResultsResolver');
 
 module.exports = {
+    isComplete,
     testResultCount,
     testResults
 };
