+++ server/resolvers/TestPlanRun/testResultCountResolver.js
@@ -1,5 +1,11 @@
 const testResultCount = testPlanRun => {
-    return testPlanRun.testResults ? testPlanRun.testResults.length : 0;
+    // check for testResult objects where 'result' exists
+    return testPlanRun.testResults
+        ? testPlanRun.testResults.reduce(
+              (acc, curr) => acc + (curr.result ? 1 : 0),
+              0
+          )
+        : 0;
 };
 
 module.exports = testResultCount;
